+++ packages/mutate-metadata/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @talismn/mutate-metadata
+
+## 0.0.2
+
+### Patch Changes
+
+- 4aa691d: feat: new balance modules